frontend: Simplify how CRs are tested
These changes use the new overrideKubeObject and a storyHelper file to
make it easy to test the CRs and CRDs.
